% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depcov.R
\docType{data}
\name{depcov}
\alias{depcov}
\title{Health insurance coverage data for young adults 2008-11}
\format{
A data frame with 150,996 rows and 41 variables:
\describe{
  \item{anyhi}{binary indicator of health insurance coverage from any source}
  \item{emphi_dep}{binary indicator of employer-sponsored health insurance coverage as a dependent}
  \item{emphi}{binary indicator of employer-sponsored health insurance coverage as an employee}
  \item{indiv}{indicator of private individual coverage}
  \item{govhi}{indicator of government-provided health insurance}
  \item{fedelig}{binary indicator of being eligible for the ACA's dependent coverage provision, i.e., age 19-25}
  \item{st_with_law}{binary indicator of a state with an existing dependent coverage law prior to the ACA}
  \item{fipstate}{FIPS code for the state}
  \item{trend}{integer counting up from 0 (Aug 2008) to 39 (Nov 2011); created by: (year-2008)*12 +(month-8)}
  \item{year}{year (2008 to 2011)}
  \item{month}{month of year (1 to 12)}
  \item{ageXX}{dummy variables for each single year of age from XX=16 to XX=29 (omitting age 26 because of ambiguity about coverage at that age)}
  \item{female}{binary indicator of female sex}
  \item{hispanic}{binary indicator of Hispanic ethnicity}
  \item{white}{binary indicator of White race}
  \item{asian}{binary indicator of Asian race}
  \item{other}{binary indicator of other race}
  \item{mar}{binary indicator of married}
  \item{student}{binary indicator of student}
  \item{fpl_ratio}{household income as a share of federal poverty line}
  \item{fpl_ratio_2}{household income as a share of federal poverty line, squared}
  \item{ue}{monthly unemployment at state level, http://127.0.0.1:21599/graphics/plot_zoom_png?width=760&height=855}
  \item{weight}{person weights from SIPP, standardized to sum to 1}
  \item{enact.trt.month}{an indicator for the period after ACA enactment but before implementation}
  \item{impl.trt.month}{an indicator for the period after ACA implementation}
  \item{young.ctrl}{binary indicator of age 16 to 17}
  \item{old.ctrl}{binary indicator of age 27 to 29}
}
}
\source{
https://www.openicpsr.org/openicpsr/project/114840/version/V1/view
}
\usage{
depcov
}
\description{
SIPP 2008 panel data on health insurance coverage of young adults originally analyzed in 
Antwi, Yaa Akosa, Asako S Moriya, and Kosali Simon. 
“Effects of Federal Policy to Insure Young Adults: Evidence from the 2010 Affordable Care Act’s Dependent-Coverage Mandate.” 
American Economic Journal: Economic Policy 5, no. 4 (November 2013): 1–28. https://doi.org/10.1257/pol.5.4.1.
This version of the data set modified for use in the NonInfParTren package by Bilinski and Hatfield.
Citation for the data set: Akosa Antwi, Yaa, Moriya, Asako S., and Simon, Kosali. Replication data for: Effects of Federal Policy to Insure Young Adults: Evidence from the 2010 Affordable Care Act’s Dependent-Coverage Mandate: AEJEPcode: data: depcov_dataset.dta. Nashville, TN: American Economic Association [publisher], 2013. Ann Arbor, MI: Inter-university Consortium for Political and Social Research [distributor], 2019-10-13. https://doi.org/10.3886/E114840V1-161781
}
\keyword{datasets}
