\name{NonInfParTren-package}
\alias{NonInfParTren-package}
\alias{NonInfParTren}
\docType{package}
\title{
\packageTitle{NonInfParTren}
}
\description{
\packageDescription{NonInfParTren}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NonInfParTren}
\packageIndices{NonInfParTren}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NonInfParTren}

Maintainer: \packageMaintainer{NonInfParTren}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
