% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonInfParTren-package.R
\docType{package}
\name{NonInfParTren-package}
\alias{NonInfParTren}
\alias{NonInfParTren-package}
\title{NonInfParTren: Test Reduced vs Expanded Model Coefficient Linear Combinations}
\description{
Fits reduced and expanded linear models and tests the difference of a linear combination of parameters from the fitted models using seemingly unrelated regression methods to account for the covariance between models, as described in "Nothing to See Here: A non-inferiority approach to parallel trends" by Alyssa Bilinski and Laura Hatfield.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laura-hatfield/NonInfParTren}
  \item Report bugs at \url{https://github.com/laura-hatfield/NonInfParTren/issues}
}

}
\keyword{internal}
